Fix a URL problem with background images
The url(…) function in CSS isn’t respecting the
assetPrefix set in the webpack.config.js. I tried
to fix this with various loaders but eventually
gave up.

This explicitly sets the background-image style
property in the DOM which can make use of the
existing #url function which does respect the
assetPrefix.